使用异步tun
